Rename in_batches and add some more tests in the specs
